refactor: mobile search drawer 레이아웃에 추가 / 스토어 합침
