+++ libs/stores/src/lib/kkshowSearchStore.tsx
@@ -7,6 +7,16 @@ export interface kkshowSearch {
   setLocalStorage(value: string[]): void;
 }
 
+export interface searchDrawer {
+  isOpen: boolean;
+  setIsOpen: (value: boolean) => void;
+}
+
+export interface SearchPopoverStore {
+  isOpen: boolean;
+  handlePopover: (value: boolean) => void;
+}
+
 export const useKkshowSearchStore = create<kkshowSearch>((set, get) => ({
   keyword: '',
   localStorage: [],
@@ -19,5 +29,21 @@ export const useKkshowSearchStore = create<kkshowSearch>((set, get) => ({
     set(() => ({
       localStorage: value,
     }));
+  },
+}));
+
+export const useSearchDrawer = create<searchDrawer>((set, get) => ({
+  isOpen: false,
+  setIsOpen: (value: boolean) => {
+    set(() => ({
+      isOpen: value,
+    }));
+  },
+}));
+
+export const useSearchPopoverStore = create<SearchPopoverStore>((set, get) => ({
+  isOpen: false,
+  handlePopover: (value) => {
+    return set({ isOpen: value });
   },
 }));
